Move header_more_buttons for export buttons to SpreadsheetExportMixin
Also remove the unnecessary call to get_header_more_buttons in WagtailAdminTemplateMixin, as the buttons are already constructed as part of get_header_buttons()
